Add validation handler
